lexer

+++ srcs/lexer/lexer.h
@@ -6,7 +6,7 @@
 /*   By: tasano <tasano@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/09 23:54:22 by tasano            #+#    #+#             */
-/*   Updated: 2022/12/28 09:27:32 by tasano           ###   ########.fr       */
+/*   Updated: 2023/01/16 00:51:44 by tasano           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,6 +45,6 @@ t_token_lst *lexer(char *line);
 t_token_lst	*token_lstnew(char *word, t_token_type type);
 void		token_lstadd_back(t_token_lst **lst, t_token_lst	*new);
 t_token_lst	*token_lstlast(t_token_lst *lst);
-size_t 		add_lst(t_token_lst **lst, char *line, size_t size, t_token_type type);
+int 		add_lst(t_token_lst **lst, char *line, size_t size, t_token_type type);
 void		token_lstfree(t_token_lst	**lst);
 #endif
